Implemented an option for delayed updating of buffers.
When editing documents it is not always useful for the file to be updated in bro
wser every keystroke, and can often lead to loading broken code into the browser
, especially when editing linked javascript and css files.

I have implemented functionality which delays the update until activity ceases f
or 1 second.   This allows you to type in a complete style rule or tag without h
aving the browser load broken code.

I have introduced the buffer local variable imp--edit-timer which stores the tim
er, and the global variable impatient-mode-delayed-update which controls whether
 updates are instantaneous or delayed.

Also I have seperated some code from imp--on-change and put it in a new function
, imp--update-pages.